Bug 1741665 - Align nsCString's public size_type better with other C++ APIs, r=mccr8,geckoview-reviewers,agi
Differential Revision: https://phabricator.services.mozilla.com/D131422

UltraBlame original commit: 43563b9a5f5d132449ec86456cf246b85ccf8306

+++ modules/libpref/Preferences.cpp
@@ -143,7 +143,7 @@ static const uint32_t MAX_ADVISABLE_PREF_LENGTH = 4 * 1024;
 
 
 static void SerializeAndAppendString(const nsCString& aChars, nsCString& aStr) {
-  aStr.AppendInt(aChars.Length());
+  aStr.AppendInt(uint64_t(aChars.Length()));
   aStr.Append('/');
   aStr.Append(aChars);
 }
